+++ src/main/java/com/atlauncher/gui/tabs/settings/SettingsTab.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package com.atlauncher.gui.tabs;
+package com.atlauncher.gui.tabs.settings;
 
 import java.awt.BorderLayout;
 import java.util.Arrays;
@@ -25,6 +25,7 @@ import javax.swing.JButton;
 import javax.swing.JPanel;
 import javax.swing.JTabbedPane;
 
+import com.atlauncher.gui.tabs.Tab;
 import org.mini2Dx.gettext.GetText;
 
 import com.atlauncher.App;
